Day 13. Distress Signal.
Did in Ruby because `eval` allows direct conversion of text to lists
without building a parser.
Rewrote comparator function so it follows the same convention as custom
sort comparator functions. This allows it to be passed directly to
Ruby's `sort` function.
Since the comparator needs to be recursive anyway, the base case is the
Integer comparison. Pull out all the type checks to the top level.